feat: Added abstract (undefined) tasks to RandomProcessGenerator
