+++ test-util/src/main/java/io/camunda/zeebe/test/util/bpmn/random/blocks/BlockSequenceBuilder.java
@@ -43,7 +43,8 @@ public class BlockSequenceBuilder extends AbstractBlockBuilder {
           new UserTaskBlockBuilder.Factory(),
           new ManualTaskBlockBuilder.Factory(),
           new IntermediateThrowEventBlockBuilder.Factory(),
-          new EscalationEventBlockBuilder.Factory());
+          new EscalationEventBlockBuilder.Factory(),
+          new AbstractBlockBuilder.Factory());
 
   private final List<BlockBuilder> blockBuilders = new ArrayList<>();
